requirements.txt
tensorflow_datasets 
akida
cnn2snn
akida-models
pillow
matplotlib
pynput